+++ src/components/DepartmentsAutocomplete.tsx
@@ -11,7 +11,7 @@ import {
 } from '@choc-ui/chakra-autocomplete';
 import { WPItem } from '@lib/classes';
 import { SearchContext } from '@context/SearchContext';
-import usePositions from '@hooks/queries/usePositions';
+import usePositions from '@queries/usePositions';
 
 export default function DepartmentsAutocomplete() {
 	const { searchDispatch } = useContext(SearchContext);