+++ src/components/EditConflictDateRangeModal.tsx
@@ -8,8 +8,8 @@ import {
 	useToast,
 } from '@chakra-ui/react';
 import { DateRange } from '@lib/classes';
-import useViewer from '@hooks/queries/useViewer';
-import useUpdateConflictRange from '@hooks/mutations/useUpdateConflictRange';
+import useViewer from '@queries/useViewer';
+import useUpdateConflictRange from '@mutations/useUpdateConflictRange';
 import { useErrorMessage } from '@hooks/hooks';
 import DateRangePicker from '@common/inputs/DateRangePicker';