+++ src/components/LoggedIn.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Box, Spinner } from '@chakra-ui/react';
 import LoginView from '@views/LoginView';
-import useViewer from '@hooks/queries/useViewer';
+import useViewer from '@queries/useViewer';
 
 interface Props {
 	hideOnly?: boolean; // Only hide the child element, don't show the login view.