+++ src/hooks/mutations/useToggleDisableProfile.tsx
@@ -3,7 +3,7 @@
  */
 
 import { gql, useMutation } from '@apollo/client';
-import { QUERY_VIEWER } from '@hooks/queries/useViewer';
+import { QUERY_VIEWER } from '@queries/useViewer';
 
 const MUTATE_TOGGLE_SEARCH_ONLY = gql`
 	mutation ToggleDisableProfile($userId: Int!) {