+++ src/views/ChangeProfileSlugView.tsx
@@ -11,8 +11,8 @@ import {
 } from '@chakra-ui/react';
 import { FiCheck, FiCopy } from 'react-icons/fi';
 import { useErrorMessage, useProfileUrl, useValidateProfileSlug } from '@hooks/hooks';
-import useViewer from '@hooks/queries/useViewer';
-import useChangeProfileSlug from '@hooks/mutations/useChangeProfileSlug';
+import useViewer from '@queries/useViewer';
+import useChangeProfileSlug from '@mutations/useChangeProfileSlug';
 import TextInput from '@common/inputs/TextInput';
 
 export default function ChangeProfileUrlView() {