+++ src/views/ResetPasswordView.tsx
@@ -16,7 +16,7 @@ import { FiHelpCircle } from 'react-icons/fi';
 import TextInput from '@common/inputs/TextInput';
 import { useErrorMessage, useValidatePassword } from '@hooks/hooks';
 import { ChangePasswordInput } from '@lib/types';
-import useResetUserPassword from '@hooks/mutations/useResetUserPassword';
+import useResetUserPassword from '@mutations/useResetUserPassword';
 
 interface Props {
 	token: string;