+++ src/views/SearchResultsView.tsx
@@ -3,7 +3,7 @@ import { Box, Flex, Link, Text } from '@chakra-ui/react';
 import { Link as RouterLink } from 'react-router-dom';
 import { FiCalendar } from 'react-icons/fi';
 import { SearchContext } from '@context/SearchContext';
-import useSavedSearches from '@hooks/queries/useSavedSearches';
+import useSavedSearches from '@queries/useSavedSearches';
 import TextCenterline from '@common/TextCenterline';
 import SavedSearchItem from '@components/SavedSearchItem';
 import CandidateList from '@components/CandidateList';